# api/openapi.yaml
openapi: 3.1.0
info:
  title: Alexandria User Service API
  description: >-
    User account management. Owns the users table; exposes internal fan-out
    endpoints consumed by knowledgebase-service and qa-service.
  license:
    name: MIT
    identifier: MIT
  version: 2.0.0
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/v1/knowledgebase/documents/{id}/tags:
    post:
      tags:
        - KnowledgeBase Service
      summary: Add tag to document
      operationId: addTag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddTagRequest'
        required: true
      responses:
        '204':
          description: Tag added
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/documents/{id}/reprocess/summary:
    post:
      tags:
        - KnowledgeBase Service
      summary: Reprocess document summary
      operationId: reprocessSummary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document summary reprocessed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/documents/{id}/reprocess/entities:
    post:
      tags:
        - KnowledgeBase Service
      summary: Reprocess document entities
      operationId: reprocessEntities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document entities reprocessed
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/documents/upload:
    post:
      tags:
        - KnowledgeBase Service
      summary: Upload a document file
      operationId: uploadDocument
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '201':
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/documents/create:
    post:
      tags:
        - KnowledgeBase Service
      summary: Create a new document with text content
      operationId: createDocument
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
        required: true
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/documents/{id}:
    get:
      tags:
        - KnowledgeBase Service
      summary: Get document by ID
      operationId: getDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
    delete:
      tags:
        - KnowledgeBase Service
      summary: Delete document
      operationId: deleteDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document deleted
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
    patch:
      tags:
        - KnowledgeBase Service
      summary: Update document metadata
      operationId: updateDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
        required: true
      responses:
        '200':
          description: Document updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
  /api/v1/knowledgebase/tags:
    get:
      tags:
        - KnowledgeBase Service
      summary: Get all unique tags for user
      operationId: getTags
      responses:
        '200':
          description: Tags retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListDto'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/search:
    get:
      tags:
        - KnowledgeBase Service
      summary: Search documents
      operationId: search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/history/search:
    get:
      tags:
        - KnowledgeBase Service
      summary: Get search history
      operationId: getSearchHistory
      responses:
        '200':
          description: Search history retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchQuery'
      security:
        - bearerAuth: []
    delete:
      tags:
        - KnowledgeBase Service
      summary: Delete search history
      operationId: deleteSearchHistory
      responses:
        '204':
          description: Search history deleted
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/documents:
    get:
      tags:
        - KnowledgeBase Service
      summary: List all documents
      operationId: listDocuments
      responses:
        '200':
          description: Documents retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/documents/{id}/summary:
    get:
      tags:
        - KnowledgeBase Service
      summary: Get document summary
      operationId: getSummary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document summary retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSummaryDto'
        '204':
          description: Summary not yet available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSummaryDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/documents/{id}/entities:
    get:
      tags:
        - KnowledgeBase Service
      summary: Get document entities
      operationId: getEntities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document entities retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentEntityResponseDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
  /api/v1/knowledgebase/documents/{id}/download:
    get:
      tags:
        - KnowledgeBase Service
      summary: Download document file
      operationId: downloadDocument
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: File content
          content:
            application/json:
              schema:
                type: string
                format: byte
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
  /api/v1/knowledgebase/documents/{documentId}/tags/{tagId}:
    delete:
      tags:
        - KnowledgeBase Service
      summary: Remove tag from document
      operationId: removeTag
      parameters:
        - name: documentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: tagId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Tag removed
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    ErrorResponse:
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: object
    AddTagRequest:
      type: object
      properties:
        label:
          type: string
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ownerSubject:
          type: string
        fileName:
          type: string
        objectKey:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        summary:
          $ref: '#/components/schemas/Summary'
        extractedEntities:
          type: array
          items:
            $ref: '#/components/schemas/ExtractedEntity'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          uniqueItems: true
    ExtractedEntity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        document:
          $ref: '#/components/schemas/Document'
        name:
          type: string
        type:
          type: string
          enum:
            - PERSON
            - DATE
            - TOPIC
            - ORGANIZATION
        confidence:
          type: number
          format: double
    Summary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        document:
          $ref: '#/components/schemas/Document'
        content:
          type: string
        generatedAt:
          type: string
          format: date-time
        modelUsed:
          type: string
    Tag:
      type: object
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        source:
          type: string
          enum:
            - AUTO
            - USER
        documents:
          type: array
          items:
            $ref: '#/components/schemas/Document'
          uniqueItems: true
    CreateDocumentRequest:
      type: object
      properties:
        fileName:
          type: string
        objectKey:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
          format: int64
        textContent:
          type: string
    UpdateDocumentRequest:
      type: object
      properties:
        fileName:
          type: string
    TagDto:
      type: object
      properties:
        name:
          type: string
        documentCount:
          type: integer
          format: int64
    TagListDto:
      type: object
      properties:
        tags:
          type: array
          items:
            $ref: '#/components/schemas/TagDto'
    SearchQuery:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userSubject:
          type: string
        queryText:
          type: string
        timestamp:
          type: string
          format: date-time
        resultCount:
          type: integer
          format: int32
    DocumentSummaryDto:
      type: object
      properties:
        summary:
          type: string
        modelUsed:
          type: string
        generatedAt:
          type: string
          format: date-time
    DocumentEntityDto:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - PERSON
            - DATE
            - TOPIC
            - ORGANIZATION
        confidence:
          type: number
          format: double
    DocumentEntityResponseDto:
      type: object
      properties:
        documentId:
          type: string
          format: uuid
        entities:
          type: array
          items:
            $ref: '#/components/schemas/DocumentEntityDto'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...